% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{am_cells_ply}
\alias{am_cells_ply}
\title{AquaMaps cells polygon Habitat Zones (`am_cells_ply`)}
\format{
A spatial features (`sf`) data frame with 1189 features and 6 fields:
\describe{
  \item{cell_id}{unique cell identifier (integer), the rownumber for the unique combination of `zone_key` and `hcaf_id`}
  \item{zone_key}{name of zone from `oh_zones`, originally BOEM Planning Area `RESA_summa` column}
  \item{hcaf_id}{half degree cell identifier (integer) from `am_cells_grd`}
  \item{CsquareCode}{from AquaMaps hcaf}
  \item{CenterLong}{from AquaMaps hcaf}
  \item{CenterLat}{from AquaMaps hcaf}
  \item{area_km2}{area of cell in square kilometers}
  \item{geom}{geometry in geographic coordinate system (EPSG: 4326)}
}
}
\source{
[Geographic Mapping Data in Digital Format](https://www.data.boem.gov/Main/Mapping.aspx),
[Marine Regions · United States Exclusive Economic Zone (EEZ)](https://marineregions.org/gazetteer.php?p=details&id=8456)
}
\usage{
am_cells_ply
}
\description{
AquaMaps hcaf cells intersected with `oh_zones`.
}
\concept{data}
\keyword{datasets}
