% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boem_blocks}
\alias{boem_blocks}
\title{BOEM Wind Energy Area Block polygons}
\format{
A spatial features (\code{sf}) data frame with 5,596 features and 24 fields:
\describe{
\item{block_id}{unique block identifier (integer)}
\item{block_key}{unique block identifier (character), comprised of
\{zone_key\}\{PROTRACTION_NUMBER\}_\{BLOCK_NUMBER\}}
\item{block_type}{one of either "lease" or "plan" depending on data source}
\item{zone_key}{unique zone from \code{oh_zones} containing the block}
\item{PROTRACTION_NUMBER}{the BOEM area defined by the roughly
2° longitude by 1° latitude containing up to 1,200 blocks.
Computed using the International UTM Zone Numbering system, e.g.
NJ is North (N); Lat 36-40° (J); 18 is UTM zone 18; and 02 is the 2nd of
12 possible maps}
\item{BLOCK_NUMBER}{Each whole block is approximately 3 nautical miles (nm)
by 3 nm. For renewable energy purposes only, blocks can be leased
down to the 16th of a block.}
\item{BLOCK_LABEL}{}
\item{SUB_BLOCK}{}
\item{lease_LEASE_NUMBER}{}
\item{lease_LEASE_TYPE}{}
\item{lease_RESOURCE}{}
\item{lease_COMPANY}{}
\item{lease_LEASE_DATE}{}
\item{lease_LEASE_TERM}{}
\item{lease_PROJECT_EASEMENT}{}
\item{lease_STATE}{}
\item{lease_PROTRACTION_NAME}{}
\item{lease_LEASE_DOCUMENT1}{}
\item{lease_LEASE_DOCUMENT2}{}
\item{plan_ADDITIONAL_INFORMATION}{}
\item{plan_CATEGORY1}{}
\item{plan_CATEGORY2}{}
\item{plan_URL1}{}
\item{plan_URL2}{}
\item{area_km2}{area of zone in square kilometers}
\item{geom}{geometry in geographic coordinate system (EPSG: 4326)}
}
}
\source{
\itemize{
\item \href{https://www.boem.gov/renewable-energy/mapping-and-data/renewable-energy-gis-data}{Renewable Energy GIS Data | Bureau of Ocean Energy Management}
}
}
\usage{
boem_blocks
}
\description{
BOEM Wind Energy Area Block polygons
}
\concept{data}
\keyword{datasets}
