% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_index}
\alias{create_index}
\title{Create index in database}
\usage{
create_index(
  con,
  tbl,
  flds,
  geom = F,
  unique = F,
  overwrite = F,
  show = F,
  exec = T
)
}
\arguments{
\item{con}{database connection from \code{DBI::dbConnect()}, eg from \code{am_pg_con()}}

\item{tbl}{table name as character}

\item{flds}{character vector of fields in table used for index}

\item{geom}{logical (default: FALSE) whether geometry field, so create GIST() index}

\item{unique}{logical (default: FALSE) whether to impose a unique constraint, to prevent duplicates; default: FALSE}

\item{overwrite}{logical (default: FALSE) whether to overwrite existing index}

\item{show}{logical (default: FALSE) whether to show SQL statement}

\item{exec}{logical (default: TRUE) whether to execute SQL statement}
}
\value{
nothing
}
\description{
Create index in database
}
\examples{
\dontrun{
con <- am_pg_con()
create_index(con, "am_cell_blocks", "geom", geom=T)
create_index(con, "am_cell_blocks", "zcb_id")
}
}
\concept{db}
