% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.R
\name{drop_na_lyrs}
\alias{drop_na_lyrs}
\title{Drop raster layers without any data}
\usage{
drop_na_lyrs(x)
}
\arguments{
\item{x}{raster layer}
}
\value{
raster stack
}
\description{
Drop \code{terra::rast()} layers without any data
}
\concept{rast}
