% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{dt_color_fld}
\alias{dt_color_fld}
\title{Colorize field in a datatable}
\usage{
dt_color_fld(dt, fld_data, fld_style = glue::glue("{fld_data}_clr"), alpha = 1)
}
\arguments{
\item{dt}{interactive \code{DT::datatable()}}

\item{alpha}{transparency (0 to 1); default = 0.5}

\item{fld}{field name in data table}
}
\value{
\code{DT::datatable()}
}
\description{
Apply a spectral color based on range of values in a field of a \code{DT::datatable()}.
}
\concept{report}
