% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud.R
\name{gcs_to_gee}
\alias{gcs_to_gee}
\title{Load into Google Earth Engine from Google Cloud Storage}
\usage{
gcs_to_gee(
  gcs_name,
  gee_name = fs::path_ext_remove(gcs_name),
  properties = "",
  gcs_bucket = "offhab_lyrs",
  gee_asset = "projects/ee-offhab/assets/lyrs_aphia_web",
  pyramiding_policy = "MEAN",
  missing_data = 255
)
}
\arguments{
\item{gcs_name}{Google Cloud Storage (GCS) file name}

\item{gee_name}{Google Earth Engine (GEE) file name (cannot have \code{.});
defaults to \code{gcs_name} without the file extension}

\item{properties}{a list of properties or blank (default: \code{""})}

\item{gcs_bucket}{GCS bucket (like a folder) for containing \code{gcs_name}; default: "offhab_lyrs"}

\item{gee_asset}{GEE asset (like a folder) for containing \code{gee_name}; default: "projects/ee-offhab/assets/lyrs_aphia_web"}

\item{pyramiding_policy}{GEE pyramiding; default: "MEAN" (but if categorical data want something else here)}

\item{missing_data}{GEE missing data; default: 255 (but need something esle if not INT1U)}
}
\value{
string of \code{{gee_asset}/{gee_name}} to reference image in GEE
}
\description{
Load into Google Earth Engine from Google Cloud Storage
}
\concept{gcs}
