% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquamaps.R
\name{get_am_cells_grd}
\alias{get_am_cells_grd}
\title{Get AquaMaps cell grid}
\usage{
get_am_cells_grd()
}
\value{
A `terra` raster layer with 1 layer:
\describe{
  \item{hcaf_id}{half degree cell identifier (integer) from `am_cells_grd`}
}
}
\description{
Get raster that only contains the `hcaf_id`.
}
\details{
Cannot simply load this as an available dataset into the R package since
`terra::rast()` only loads a pointer to raster file, not its contents.
}
\examples{
am_cells_grd <- get_am_cells_grd()
}
\concept{aquamaps}
