% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_colors}
\alias{get_colors}
\title{Get Colors}
\usage{
get_colors(x, v_colors = oh_colors)
}
\arguments{
\item{x}{character vector of names}

\item{v_colors}{vector of named characters with colors; defaults to \code{oh_colors}}
}
\value{
hexadecimal colors
}
\description{
typically used with \code{leaflet::addPolygons()}
}
\concept{viz}
