% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{link_caption}
\alias{link_caption}
\title{Add extra link to caption}
\usage{
link_caption(
  txt,
  txt_link = ". Please also see the [online interactive figure]",
  condition_skip = knitr::is_html_output(),
  url_pfx = "https://offshorewindhabitat.info/report"
)
}
\arguments{
\item{txt}{text of caption to always include}

\item{txt_link}{text to additionally include if condition_skip not TRUE}

\item{condition_skip}{condition of when to skip including extra \code{txt_link}}

\item{url_pfx}{link to prefix the output}
}
\value{
text
}
\description{
Add link to caption for static reports (i.e. docx or pdf).
}
\concept{report}
