% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oa_regions}
\alias{oa_regions}
\title{OceanAdapt regions for bottom trawl data}
\format{
A spatial features (\code{sf}) data frame with 7 features and 4 fields:
\describe{
\item{oa_rgn}{unique 3-letter lowercase text key}
\item{oa_region}{name of region}
\item{oa_region_rds}{name of region used in the stored data file (*.rds)}
\item{geometry}{}
\item{oa_region}{}
\item{oa_region_rds}{}
\item{geometry}{geometry in geographic coordinate system (EPSG: 4326)}
}
}
\source{
\href{https://www.fisheries.noaa.gov/inport/item/67352}{DisMAP Regions 20220516 | InPort}
}
\usage{
oa_regions
}
\description{
Polygons define the regions used in the interpolated bottom trawl data.
}
\concept{data}
\keyword{datasets}
