% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oh_colors}
\alias{oh_colors}
\title{OffHab colors}
\format{
An object of class \code{character} of length 14.
}
\usage{
oh_colors
}
\description{
vector of named character colors corresponding with \code{region} or \code{zone_name}
from \code{oh_zones}
}
\concept{data}
\keyword{datasets}
