% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{oh_con}
\alias{oh_con}
\title{Connect to OffHab database}
\usage{
oh_con(
  path_db = system.file("offhab.duckdb", package = "offhabr"),
  read_only = T
)
}
\arguments{
\item{path_db}{path to database file, defaults to duckdb in this R package}

\item{read_only}{only applicable to duckdb (not SQLite); defaults to TRUE; vs set to FALSE if writing to database}
}
\value{
a \code{DBI::dbConnect()} object
}
\description{
Using local filesystem database \code{RSQLite::SQLite()}
}
\examples{
\dontrun{
con <- oh_con()
DBI::dbListTables(con)
}
}
\concept{db}
