% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{oh_map}
\alias{oh_map}
\title{Setup interactive map with OffHab defaults}
\usage{
oh_map(
  base_opacity = 0.5,
  zoomControl = knitr::is_html_output(),
  attributionControl = knitr::is_html_output(),
  fullScreenControl = knitr::is_html_output(),
  verbose = F
)
}
\arguments{
\item{base_opacity}{the opacity of the basemap; default is 0.5}

\item{zoomControl}{boolean (TRUE or FALSE) for \code{leaflet::leafletOptions()}
on whether to include zoom; default: \code{TRUE}}

\item{attributionControl}{boolean (TRUE or FALSE) for \code{leaflet::leafletOptions()}
on whether to include attribution; default: \code{TRUE}}

\item{fullScreenControl}{boolean (TRUE or FALSE) on whether to include
\code{leaflet.extras::addFullscreenControl()}; default: \code{TRUE}}

\item{verbose}{boolean (TRUE or FALSE) on whether to show debug information; default: \code{FALSE}}
}
\value{
a \code{leaflet::leaflet()} map
}
\description{
Make an interactive map with basemap
}
\examples{
# basic map with defaults
oh_map()

# map zones by area with divergent color scheme around mean
oh_map(
  ply     = oh_zones_s1k,
  fld_val = area_km2,
  fld_id  = zone_name,
  str_val = "area (km^2)",
  str_id  = "Zone",
  div_mid = mean(oh_zones$area_km2))
}
\concept{viz}
