% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{oh_map_cog_lyr}
\alias{oh_map_cog_lyr}
\title{Map Cloud-Optimize TIFF OffHab layer}
\usage{
oh_map_cog_lyr(
  lyr_key,
  lyr_title = "\% Habitat",
  con = con,
  show_zones = T,
  ...
)
}
\arguments{
\item{lyr_key}{layer key (in form of \verb{\{dataset_key\}_\{aphia_id\}})}

\item{lyr_title}{layer title}

\item{con}{database connection}

\item{...}{parameters passed onto \code{oh_map_cog()}}
}
\value{
\code{leaflet::leaflet()} object
}
\description{
Use offhabr \code{lyr_key} to pull range from database and render map with \code{oh_map_cog()}
}
\concept{viz}
