% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{oh_map_cog_sp}
\alias{oh_map_cog_sp}
\title{Map Cloud-Optimize TIFF OffHab Species}
\usage{
oh_map_cog_sp(aphia_id, lyr_title = "\% Habitat", con = con, ...)
}
\arguments{
\item{...}{}
}
\description{
Use offhabr \code{aphia_id} to pull range from database and render map with \code{oh_map_cog()}
}
\concept{viz}
