% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{oh_map_ply}
\alias{oh_map_ply}
\title{Map polygons}
\usage{
oh_map_ply(
  ply,
  fld_val,
  fld_id,
  str_val,
  str_id,
  div_mid = NULL,
  col_pal = "Spectral",
  fill_opacity = 0.6,
  brdr_color = "gray",
  brdr_opacity = 0.9,
  brdr_weight = 1,
  base_map = "CartoDB.Positron",
  base_opacity = 0.5
)
}
\arguments{
\item{ply}{a `sf` polygon data frame}

\item{fld_val}{an unquoted field name (see `tidy-select`) for the column in
`ply` containing the value to map}

\item{fld_id}{an unquoted field name (see `tidy-select`) for the column in
`ply` containing the identifier to include in the popup}

\item{str_val}{the string name for the field value in the legend and popup}

\item{str_id}{the string name for the identifier in the popup}

\item{div_mid}{the divergent midpoint of the field value `fld_val` for
setting as the midpoint of the color palette `col_pal`; default is `NULL`}

\item{col_pal}{color palette for the polygon fill color; default is
"Spectral"}

\item{fill_opacity}{fill opacity; default is 0.6}

\item{brdr_color}{border color; default is "gray"}

\item{brdr_opacity}{border opacity; default is 0.9}

\item{brdr_weight}{border weight; default is 1}

\item{base_map}{the basemap (see  `leaflet$providers`); default is
"CartoDB.Positron"}

\item{base_opacity}{the opacity of the basemap; default is 0.5}
}
\value{
a `leaflet::leaflet()` map
}
\description{
Make an interactive choropleth map, optionally with a color scheme that
diverges around a midpoint value.
}
\examples{
# map zones by area
oh_map_ply(
  ply     = oh_zones_s1k,
  fld_val = area_km2,
  fld_id  = zone_name,
  str_val = "area (km^2)",
  str_id  = "Zone")

# map zones by area with divergent color scheme around mean
oh_map_ply(
  ply     = oh_zones_s1k,
  fld_val = area_km2,
  fld_id  = zone_name,
  str_val = "area (km^2)",
  str_id  = "Zone",
  div_mid = mean(oh_zones$area_km2))
}
\concept{viz}
