% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{oh_map_zone_score_dev}
\alias{oh_map_zone_score_dev}
\title{Map score as deviation from average in zone with blocks}
\usage{
oh_map_zone_score_dev(
  zone_key,
  stk_web_tif = "~/My Drive/projects/offhab/data/derived/stack_web.tif",
  zonal_blocks_csv = "~/Github/ecoquants/offhab-scripts/data/zonal_blocks.csv"
)
}
\arguments{
\item{zone_key}{one of \code{zone_key} from \code{offhabr::oh_zones}}

\item{stk_web_tif}{path to \code{stack_web.tif} containing layers "oh_zones_v1_web" and "score_v1_web"}

\item{zonal_blocks_csv}{path to zonal average values of blocks}
}
\value{
htmlwidget map as \code{leaflet::leaflet()} with \code{attr("zone_name")}
}
\description{
Map score as deviation from average in zone with blocks
}
