% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oh_study_v1}
\alias{oh_study_v1}
\title{Offshore Habitat study area (version 1)}
\format{
An object of class \code{sfc_MULTIPOLYGON} (inherits from \code{sfc}) of length 1.
}
\usage{
oh_study_v1
}
\description{
Offshore habitat zones \code{oh_zones} (filtered by \code{zone_version}==1) dissolved into a single study area polygon.
}
\concept{data}
\keyword{datasets}
