% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{oh_tbl_zone_score_dev}
\alias{oh_tbl_zone_score_dev}
\title{Tabulate Blocks with Score as deviation from average in zone}
\usage{
oh_tbl_zone_score_dev(m)
}
\arguments{
\item{m}{htmlwidget map as \code{leaflet::leaflet()} with attributes "zone_name",
"zone_key" and "block_data" as returned from \code{oh_map_zone_score_dev()}}
}
\value{
either a table of type \code{DT::datatable()} for interactive html output
(if \code{knitr::is_html_output() == TRUE}), or a static table of type \code{gt:gt()}
for Word docx output
}
\description{
Tabulate Blocks with Score as deviation from average in zone
}
\concept{report}
