% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oh_zones_s1k}
\alias{oh_zones_s1k}
\title{Offshore Habitat Zone polygons, simplified to 1 km}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 22 rows and 11 columns.
}
\usage{
oh_zones_s1k
}
\description{
Same as \code{oh_zones}, simplified to 1 km for faster rendering of smaller output files.
}
\seealso{
\link{oh_zones}
}
\concept{data}
\keyword{datasets}
