% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{rl_pg_con}
\alias{rl_pg_con}
\title{Connect to IUCN RedList Postgres database}
\usage{
rl_pg_con()
}
\value{
a \code{DBI::dbConnect()} object
}
\description{
Connect to IUCN RedList Postgres database
}
\examples{
\dontrun{
con <- rl_pg_con()
DBI::dbListTables(con)
}
}
\concept{db}
