% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud.R
\name{upload_to_gcs}
\alias{upload_to_gcs}
\title{Upload to Google Cloud Storage}
\usage{
upload_to_gcs(
  file,
  name = basename(file),
  make_public = T,
  gcs_auth_json =
    "/Users/bbest/My Drive/private/offhab-google-service-account_09e7228ac965.json",
  gcs_bucket = "offhab_lyrs"
)
}
\arguments{
\item{file}{usually a GeoTIFF}

\item{name}{defaults to \code{basename(file)}}

\item{make_public}{defaults to True}

\item{gcs_auth_json}{JSON for authorizing}

\item{gcs_bucket}{bucket (i.e. folder) to upload into on Google Cloud Storage}
}
\value{
empty, only side effect
}
\description{
Upload to Google Cloud Storage
}
\concept{gcs}
