% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worms.R
\name{wm_add_aphia_id}
\alias{wm_add_aphia_id}
\title{Match taxa to WoRMS aphia_id and update database}
\usage{
wm_add_aphia_id(
  df,
  fld,
  tbl_str = deparse(substitute(df)),
  fld_str = deparse(substitute(fld))
)
}
\arguments{
\item{df}{data frame}

\item{fld}{unquoted field name containing taxonomic field, eg `taxa`}

\item{tbl_str}{the source table to be written into the database referenced in `taxa.tbl`}

\item{fld_str}{the source field to be written into the database referenced in `taxa.fld`}
}
\value{
the original data frame `df` with `aphia_id` column prepended
}
\description{
And optionally append to taxa* tables in database
}
\examples{
\dontrun{
tmp_test <- tibble::tribble(
        ~common,                  ~scientific,
  "Minke whale",  "Balaenoptera acutorostrata",
   "Blue whale",  "Balaenoptera musculus")
wm_add_aphia_id(df_test, scientific, tbl_str = "tmp_test")
}
}
