% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wm_rest_params}
\alias{wm_rest_params}
\title{Lookup table for maximum requests per WoRMS REST API used by \code{wm_rest()}}
\format{
A data frame with 3 rows and 3 columns:
\describe{
\item{type}{helpful descriptor of type of operation}
\item{operation}{name of operation}
\item{max}{maximum number of simultaneous requests allowed}
}
}
\source{
\href{https://www.marinespecies.org/rest/}{WoRMS REST API}
}
\usage{
wm_rest_params
}
\description{
A data frame describing the maximum number of simultaneous requests allowed
by the \href{https://www.marinespecies.org/rest/}{WoRMS REST API}.
}
\concept{data}
\keyword{datasets}
