% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{am_cell_blocks}
\alias{am_cell_blocks}
\title{AquaMaps cell polygons intersecting OffHab zones and BOEM blocks}
\format{
A spatial features (`sf`) data frame with 1189 features and 6 fields:
\describe{
  \item{zcb_id}{unique identifier (integer) for OffHab zone `oh_zones` x
    AquaMaps cell `am_cells` x BOEM block `boem_blocks`}
  \item{block_key}{unique block identifier (character), comprised of
    \{PROTRACTION_NUMBER\}_\{BLOCK_NUMBER\}}
  \item{block_type}{one of either "lease" or "plan" depending on data source}
  \item{zone_key}{name of zone from `oh_zones`, originally BOEM Planning Area `RESA_summa` column}
  \item{zc_id}{identifier (integer) for OffHab zone and AquaMaps cell}
  \item{hcaf_id}{AquaMaps half degree cell identifier (integer) from `am_cells_grd`}
  \item{csquare_code}{from AquaMaps hcaf}
  \item{area_km2}{area of cell in square kilometers}
  \item{geom}{geometry in geographic coordinate system (EPSG: 4326)}
}
}
\source{
[Geographic Mapping Data in Digital Format](https://www.data.boem.gov/Main/Mapping.aspx),
[Marine Regions · United States Exclusive Economic Zone (EEZ)](https://marineregions.org/gazetteer.php?p=details&id=8456)
}
\usage{
am_cell_blocks
}
\description{
Polygons from intersecting OffHab zone `oh_zones` x AquaMaps cell `am_cells` x BOEM block
`boem_blocks`
}
\concept{data}
\keyword{datasets}
