% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{am_pg_con}
\alias{am_pg_con}
\title{Connect to AquaMaps Postgres database}
\usage{
am_pg_con()
}
\value{
a \code{DBI::dbConnect()} object
}
\description{
Connect to AquaMaps Postgres database
}
\examples{
\dontrun{
con <- am_pg_con()
DBI::dbListTables(con)
}
}
\concept{db}
