% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquamaps.R
\name{get_am_cell_zones_grd}
\alias{get_am_cell_zones_grd}
\title{AquaMaps cell raster contained by OffHab zones}
\usage{
get_am_cell_zones_grd()
}
\value{
A \code{terra} raster layer with 1 layer:
\describe{
\item{hcaf_id}{half degree cell identifier (integer) from \code{am_cell_zones}}
}
}
\description{
Get raster with AquaMaps \code{hcaf_id} for cells contained by OffHab zones \code{oh_zones}.
}
\details{
Cannot simply load this as an available dataset into the R package since
\code{terra::rast()} only loads a pointer to raster file, not its contents.
}
\examples{
am_cell_zones_grd <- get_am_cell_zones_grd()
}
\concept{aquamaps}
