% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{oh_con}
\alias{oh_con}
\title{Connect to OffHab database}
\usage{
oh_con(
  path_duckdb = file.path(system.file(package = "offhabr"), "offhab.duckdb"),
  read_only = T
)
}
\arguments{
\item{read_only}{defaults to FALSE, i.e. write}

\item{path_sqlite}{path to SQLite database file}
}
\value{
a \code{DBI::dbConnect()} object
}
\description{
Using local filesystem database \code{RSQLite::SQLite()}
}
\examples{
\dontrun{
con <- oh_con()
DBI::dbListTables(con)
}
}
\concept{db}
