% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{oh_map}
\alias{oh_map}
\title{Setup interactive map with OffHab defaults}
\usage{
oh_map(base_opacity = 0.5)
}
\arguments{
\item{base_opacity}{the opacity of the basemap; default is 0.5}
}
\value{
a \code{leaflet::leaflet()} map
}
\description{
Make an interactive map with basemap
}
\examples{
# basic map with defaults
oh_map()

# map zones by area with divergent color scheme around mean
oh_map(
  ply     = oh_zones_s1k,
  fld_val = area_km2,
  fld_id  = zone_name,
  str_val = "area (km^2)",
  str_id  = "Zone",
  div_mid = mean(oh_zones$area_km2))
}
\concept{viz}
