% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{oh_map_cog}
\alias{oh_map_cog}
\title{Map Cloud-Optimized GeoTIFF}
\usage{
oh_map_cog(
  cog_file,
  cog_dir = "https://storage.googleapis.com/offhab_lyrs",
  cog_range = c(1, 100),
  cog_method = "average",
  cog_colors = "viridis",
  bb = c(-129.4, 23.2, -64.7, 48.9),
  title = "\% Habitat",
  base_opacity = 0.5
)
}
\arguments{
\item{cog_file}{filename of Cloud-Optimized GeoTIFF (COG)}

\item{cog_dir}{directory of path to COG; defaults to "https://storage.googleapis.com/offhab_lyrs"}

\item{cog_range}{range of values to colorramp in COG; default: \code{c(1, 100)}}

\item{cog_method}{method for interpolation between zoom levels; choose "nearest" for categorical; default: "average" (for continuous)}

\item{cog_colors}{color ramp that applies to both \code{leaflet::colorNumeric()}
and the \code{colormap_name} of
\href{https://api.cogeo.xyz/docs#/Cloud\%20Optimized\%20GeoTIFF/tile_cog_tiles__TileMatrixSetId___z___x___y___scale_x__format__get}{https://api.cogeo.xyz/cog/tiles};
default: "viridis"}

\item{bb}{bounding box to feed \code{leaflet::fitBounds()} defaults to US48:
\code{c(-129.4, 23.2, -64.7, 48.9)}}

\item{title}{title for legend; default: \code{"\% Habitat"}}

\item{base_opacity}{the opacity of the basemap; default is \code{0.5}}
}
\value{
a \code{leaflet::leaflet()} map
}
\description{
Make an interactive map with basemap
}
\examples{
# map by aphia_id
oh_map_cog("aphia_100599_web.tif")
}
\concept{viz}
