% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{oh_pg_con}
\alias{oh_pg_con}
\title{Connect to OffHab Postgres Database}
\usage{
oh_pg_con()
}
\value{
a `DBI::dbConnect()` object
}
\description{
Connect to OffHab Postgres Database
}
\examples{
\dontrun{
con <- oh_pg_con()
DBI::dbListTables(con)
}
}
\concept{db}
