% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.R
\name{oh_rast}
\alias{oh_rast}
\title{OffHab reference raster}
\usage{
oh_rast(
  type = c("NA", "cell_id", "region", "zone_id", "block_id", "area_m2"),
  zone_id = "ALL",
  zone_version = c(1, 2)
)
}
\arguments{
\item{type}{the type of raster, being one of: \code{"NA"} (default all \code{NA} values)
\code{"cell_id"} with unique cell indices to set values of reference raster,
\code{"region"} showing the 3 unique regions ("Atlantic","Gulf of Mexico" and "Pacific"),
\code{"zone_id"} corresponding with the \code{zone_id} of \code{oh_zones},
\code{"block_id"} corresponding with the \code{block_id} of \code{oh_blocks}, or \code{"area_m2"}
for square meter area per cell}

\item{zone_id}{the \code{zone_id} (integer) to trim the output raster, being one of the
zones found in the \code{zone_id} field of \code{oh_zones} or \code{"ALL"} zones (the default)}

\item{zone_version}{the \code{zone_version} (integer) to choose, whether original BOEM
planning area clipped out to EEZ (\verb{=1}) or more restricted to OceanAdapt regions
for bottom trawl data (\verb{=2}); applies to either \code{type} of \code{"zone_id"} or \code{"block_id"}}
}
\value{
return a reference raster (from \code{terra::rast()})
}
\description{
Get OffHab reference raster of cells by \code{cell_id}, \code{zone_id} or all \code{NA} and
optionally trim to \code{zone_id}.
}
\details{
The OffHab reference raster is based on the GEBCO global bathymetry clipped to the
United States in the web Mercator projection (EPSG:3857) for readily using
with interactive maps (e.g. \code{leaflet::leaflet()}).
}
\examples{

# reference raster with all NA values
r_na <- oh_rast()
r_na

# zone_id for all zones
r_zid <- oh_rast("zone_id")
r_zid
terra::plot(r_zid)

# cell_id for all zones
r_cid <- oh_rast("cell_id")
r_cid
terra::plot(r_cid)

# cell_id for Straits of Florida (zone_id: 4; zone_key: fls)
r_cid_fls <- oh_rast("cell_id", 4)
r_cid_fls
terra::plot(r_cid_fls)
}
\concept{rast}
