% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oh_zones}
\alias{oh_zones}
\title{Offshore Habitat Zone polygons from BOEM Program Areas clipped to US EEZ}
\format{
A spatial features (\code{sf}) data frame with 10 rows and 5 columns:
\describe{
\item{zone_id}{unique zone identifier (integer)}
\item{zone_key}{unique zone identifier (character), the lower-case of original BOEM Planning Area \code{MMS_PLAN_A} column}
\item{zone_name}{name of zone, originally BOEM Planning Area \code{RESA_summa} column}
\item{region}{oceanic region; one of: "Atlantic", "Gulf of Mexico" or "Pacific"}
\item{area_km2}{area of zone in square kilometers}
\item{geom}{geometry in geographic coordinate system (EPSG: 4326)}
}
}
\source{
\href{https://www.data.boem.gov/Main/Mapping.aspx}{Geographic Mapping Data in Digital Format},
\href{https://marineregions.org/gazetteer.php?p=details&id=8456}{Marine Regions · United States Exclusive Economic Zone (EEZ)}
}
\usage{
oh_zones
}
\description{
Offshore habitat zones used for summary statistics and derived from BOEM
2019-2024 Draft Proposed Program Areas for the lower 48 (excluding Hawaii and
Alaska), clipped by the United States Exclusive Economic Zone (EEZ).
}
\concept{data}
\keyword{datasets}
