% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.R
\name{write_rast}
\alias{write_rast}
\title{Write compact raster with compression, tiles and sparse encoding}
\usage{
write_rast(r, tif, datatype = "INT1U", overwrite = TRUE, epsg = 3857)
}
\arguments{
\item{r}{input raster layer}

\item{tif}{output GeoTIFF}

\item{datatype}{datatype; one of: INT1U (default), INT2S, INT2U, INT4S, INT4U, FLT4S, FLT8S}

\item{overwrite}{defaults to TRUE}

\item{epsg}{projection number; default: Web Mercator (3857)}
}
\value{
returns nothing since only writing \code{tif}, i.e. a side-effect function
}
\description{
Use \code{terra::WriteRaster()} followed by \code{gdal_translate} for minimizing the output
file size with creation options for maximum compression
(\verb{COMPRESS=DEFLATE ZLEVEL=9 PREDICTOR=2}), tiles (\code{TILED=YES}) and sparse
(\code{SPARSE_OK=TRUE}) encoding.
}
\details{
The \code{datatype} corresponds with the following:\tabular{lrrl}{
   terra \tab min \tab max \tab gdal \cr
   INT1U \tab 0 \tab 255 \tab Byte \cr
   INT2S \tab -32,767 \tab 32,767 \tab Int16 \cr
   INT2U \tab 0 \tab 65,534 \tab UInt16 \cr
   INT4S \tab -2,147,483,647 \tab 2,147,483,647 \tab Int32 \cr
   INT4U \tab 0 \tab 4,294,967,296 \tab UInt32 \cr
   FLT4S \tab -3.4e+38 \tab 3.4e+38 \tab Float32 \cr
   FLT8S \tab -1.7e+308 \tab 1.7e+308 \tab Float64 \cr
}
}
\concept{rast}
