% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.R
\name{write_rast}
\alias{write_rast}
\title{Write compact raster with compression, tiles and sparse encoding}
\usage{
write_rast(
  r,
  tif,
  datatype = "INT1U",
  overwrite = TRUE,
  method = c("average", "nearest"),
  threads = "ALL_CPUS",
  epsg = 3857,
  verbose = F,
  web_optimize = F,
  use_gdal_cog = F
)
}
\arguments{
\item{r}{input raster layer or path to input GeoTIFF}

\item{tif}{output path to GeoTIFF}

\item{datatype}{datatype; one of: INT1U (default), INT2S, INT2U, INT4S, INT4U, FLT4S, FLT8S}

\item{overwrite}{defaults to TRUE}

\item{method}{for internal overview generation; use \code{"average"} (default) for continous
and \code{"nearest"} for categorical}

\item{threads}{used by \href{https://cogeotiff.github.io/rio-cogeo/CLI/}{\verb{rio cogeo create}}; defaults to \code{"ALL_CPUS"}; change to \code{1} if using in parallel}

\item{epsg}{projection number; default: Web Mercator (3857)}

\item{verbose}{show commands; defaults to FALSE}

\item{web_optimize}{create a web-tiling friendly COG per
\href{https://cogeotiff.github.io/rio-cogeo/Advanced/}{Advanced Topics - rio-cogeo}
that aligns internal tiles with the web-mercator grid resulting in a larger}

\item{use_gdal_cog}{use GDAL COG driver (requires GDAL >= 3.1); does not change extent
file with different bounds and pixel sizes from original; defaults to FALSE}
}
\value{
returns nothing since only writing \code{tif}, i.e. a side-effect function
}
\description{
Use \code{terra::WriteRaster()} followed by
\href{https://cogeotiff.github.io/rio-cogeo/CLI/}{\verb{rio cogeo create}} for
creating cloud-optimized GeoTIFF that minimizes the output file size with
creation options for maximum compression
(\verb{COMPRESS=DEFLATE ZLEVEL=9 PREDICTOR=2}), tiles (\code{TILED=YES}) (and
optionally sparse (\code{SPARSE_OK=TRUE}) encoding).
}
\details{
The \code{datatype} corresponds with the following:\tabular{lrrll}{
   terra \tab min \tab max \tab gdal \tab rio \cr
   INT1U \tab 0 \tab 255 \tab Byte \tab uint8 \cr
   INT2S \tab -32,767 \tab 32,767 \tab Int16 \tab int16 \cr
   INT2U \tab 0 \tab 65,534 \tab UInt16 \tab uint16 \cr
   INT4S \tab -2,147,483,647 \tab 2,147,483,647 \tab Int32 \tab int32 \cr
   INT4U \tab 0 \tab 4,294,967,296 \tab UInt32 \tab uint32 \cr
   FLT4S \tab -3.4e+38 \tab 3.4e+38 \tab Float32 \tab float32 \cr
   FLT8S \tab -1.7e+308 \tab 1.7e+308 \tab Float64 \tab float64 \cr
}
}
\concept{rast}
